requirements.txt
python-dotenv
openai
pydantic
spellchecker
langdetect
tkinter